requirements.txt
beautifulsoup4
python-docx
tqdm
requests